backend/app/services/auth_service.py
from datetime import datetime, timedelta, timezone

from jose import JWTError, jwt
from passlib.context import CryptContext

from app.config import settings

# Bcrypt hasher — intentionally slow to prevent brute force attacks
# "deprecated=auto" for if i change algorithms later,
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
def hash_password(plain_password: str) -> str:
    return pwd_context.hash(plain_password)

def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)

def create_access_token(user_id: str) -> str:
    # Short-lived token (30 min) sent with every API request
    # Contains the user's ID so the backend knows who's calling
    expire = datetime.now(timezone.utc) + timedelta(
        minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES
    )
    payload = {
        "sub": user_id,       # "sub" = subject (standard JWT claim)
        "exp": expire,        # "exp" = expiry time
        "type": "access",
    }
    return jwt.encode(payload, settings.SECRET_KEY, algorithm=settings.ALGORITHM)

def create_refresh_token(user_id: str) -> str:
    # Long-lived token (7 days) used to get new access tokens without logging in again
    expire = datetime.now(timezone.utc) + timedelta(
        days=settings.REFRESH_TOKEN_EXPIRE_DAYS
    )
    payload = {
        "sub": user_id,
        "exp": expire,
        "type": "refresh",
    }
    return jwt.encode(payload, settings.SECRET_KEY, algorithm=settings.ALGORITHM)


def verify_token(token: str, expected_type: str = "access") -> str | None:
    # Decodes a token and returns the user_id
    # but also returns None if expired, malformed, or wrong type
    try:
        payload = jwt.decode(
            token,
            settings.SECRET_KEY,
            algorithms=[settings.ALGORITHM],
        )
        user_id: str = payload.get("sub")
        token_type: str = payload.get("type")

        if user_id is None or token_type != expected_type:
            return None
        return user_id

    except JWTError:
        return None