backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
anthropic==0.39.0
openai==1.58.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.12
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
greenlet
